cpplib.c (_cpp_init_stacks): Cast enum for comparison.
	* cpplib.c (_cpp_init_stacks): Cast enum for comparison.
	* cppexp.c (lex): Cast enums for comparison.
	* cppinit.c (parse_option): Cast enum for comparison.
	* cpplex.c (cpp_spell_token): Cast enums to int for minus.
	(cpp_output_token): Likewise.
	(cpp_can_paste): Cast enums for comparsion and plus/minus.
	(cpp_avoid_paste): Cast enums for minus and comparison.

From-SVN: r40174

+++ gcc/cppinit.c
@@ -1191,7 +1191,7 @@ parse_option (input)
 		 Otherwise, return the longest option-accepting match.
 		 This loops no more than twice with current options.  */
 	      mx = md;
-	      for (; mn < N_OPTS; mn++)
+	      for (; mn < (unsigned int) N_OPTS; mn++)
 		{
 		  opt_len = cl_options[mn].opt_len;
 		  if (memcmp (input, cl_options[mn].opt_text, opt_len))
